+++ packages/backend/src/managers/catalogManager.ts
@@ -30,20 +30,9 @@ import { Publisher } from '../utils/Publisher';
 import type { LocalModelImportInfo } from '@shared/src/models/ILocalModelInfo';
 import { InferenceType } from '@shared/src/models/IInference';
 import { CatalogFormat, hasCatalogWrongFormat, merge, sanitize } from '../utils/catalogUtils';
+import type { FilterRecipesResult, RecipeFilters } from '@shared/src/models/FilterRecipesResult';
 
 export const USER_CATALOG = 'user-catalog.json';
-
-export type CatalogFilterKey = 'languages' | 'tools' | 'frameworks';
-
-export type RecipeFilters = {
-  [key in CatalogFilterKey]?: string[];
-};
-
-export interface FilterRecipesResult {
-  filters: RecipeFilters;
-  choices: RecipeFilters;
-  result: Recipe[];
-}
 
 export class CatalogManager extends Publisher<ApplicationCatalog> implements Disposable {
   private readonly _onUpdate = new EventEmitter<ApplicationCatalog>();
@@ -330,7 +319,11 @@ export class CatalogManager extends Publisher<ApplicationCatalog> implements Dis
       delete subfilters.tools;
       choices.tools = this.filterRecipes(subfilters).choices.tools;
     } else {
-      choices.tools = result.map(r => r.backend).filter(b => b !== undefined);
+      choices.tools = result
+        .map(r => r.backend)
+        .filter(b => b !== undefined)
+        .filter((value, index, array) => array.indexOf(value) === index)
+        .sort((a, b) => a.localeCompare(b));
     }
 
     if ('frameworks' in filters) {
